TS-3033: generic message initial marshalling
Implement a very simple generic message marshalling API. This is
similar to the existing hand-rolled marshalling, except that there
are fewer data types, and it is all implemented in a single place.

Add MgmtSocket.cc because there's really no benefit to having all
the mgmt socket wrappers inline. Move some more helper API over to
MgmtSocket.

+++ lib/ts/ink_inet.cc
@@ -499,13 +499,13 @@ REGRESSION_TEST(Ink_Inet) (RegressionTest * t, int /* atype */, int * pstatus) {
       box.check(ats_ip_parse(ts::ConstBuffer(names[i].hostspec, strlen(names[i].hostspec)), &addr, &port) == 0,
           "ats_ip_parse(%s)", names[i].hostspec);
       box.check(strncmp(addr.data(), names[i].host, addr.size()) ==  0,
-          "ats_ip_parse(%s) gave addr '%.*s'", names[i].hostspec, addr.size(), addr.data());
+          "ats_ip_parse(%s) gave addr '%.*s'", names[i].hostspec, (int)addr.size(), addr.data());
       if (names[i].port) {
         box.check(strncmp(port.data(), names[i].port, port.size()) ==  0,
-          "ats_ip_parse(%s) gave port '%.*s'", names[i].hostspec, port.size(), port.data());
+          "ats_ip_parse(%s) gave port '%.*s'", names[i].hostspec, (int)port.size(), port.data());
       } else {
         box.check(port.size() == 0,
-          "ats_ip_parse(%s) gave port '%.*s'", names[i].hostspec, port.size(), port.data());
+          "ats_ip_parse(%s) gave port '%.*s'", names[i].hostspec, (int)port.size(), port.data());
       }
 
     }
